Move product dialog paths to qs
