+++ saleor/static/dashboard-next/products/urls.ts
@@ -1,5 +1,7 @@
 import { stringify as stringifyQs } from "qs";
 import * as urlJoin from "url-join";
+
+import { Dialog } from "../types";
 import { ProductListQueryParams } from "./views/ProductList";
 
 const productSection = "/products/";
@@ -17,15 +19,26 @@ export const productListUrl = (params?: ProductListQueryParams): string => {
 };
 
 export const productPath = (id: string) => urlJoin(productSection, id);
-export const productUrl = (id: string) => productPath(encodeURIComponent(id));
+export type ProductUrlDialog = "remove";
+export type ProductUrlQueryParams = Dialog<"remove">;
+export const productUrl = (id: string, params?: ProductUrlQueryParams) =>
+  productPath(encodeURIComponent(id)) + "?" + stringifyQs(params);
 
 export const productVariantEditPath = (productId: string, variantId: string) =>
   urlJoin(productSection, productId, "variant", variantId);
-export const productVariantEditUrl = (productId: string, variantId: string) =>
+export type ProductVariantEditUrlDialog = "remove";
+export type ProductVariantEditUrlQueryParams = Dialog<"remove">;
+export const productVariantEditUrl = (
+  productId: string,
+  variantId: string,
+  params?: ProductVariantEditUrlQueryParams
+) =>
   productVariantEditPath(
     encodeURIComponent(productId),
     encodeURIComponent(variantId)
-  );
+  ) +
+  "?" +
+  stringifyQs(params);
 
 export const productVariantAddPath = (productId: string) =>
   urlJoin(productSection, productId, "variant/add");
@@ -34,5 +47,13 @@ export const productVariantAddUrl = (productId: string) =>
 
 export const productImagePath = (productId: string, imageId: string) =>
   urlJoin(productSection, productId, "image", imageId);
-export const productImageUrl = (productId: string, imageId: string) =>
-  productImagePath(encodeURIComponent(productId), encodeURIComponent(imageId));
+export type ProductImageUrlDialog = "remove";
+export type ProductImageUrlQueryParams = Dialog<"remove">;
+export const productImageUrl = (
+  productId: string,
+  imageId: string,
+  params?: ProductImageUrlQueryParams
+) =>
+  productImagePath(encodeURIComponent(productId), encodeURIComponent(imageId)) +
+  "?" +
+  stringifyQs(params);
